fix: humidity and distance sensors lost their icon and their label
Gladys resolves a device feature's icon AND its translated label by the
exact (category, type) pair — `DeviceFeatureCategoriesIcon[category][type]`
in the front, falling back to a generic slider, and
`deviceFeatureCategory.<category>.<type>` in the translations, falling
back to the raw feature name. A pair Gladys does not register therefore
renders as a nameless, iconless line even though the value arrives
perfectly.

`humidity-sensor` is registered with `decimal` alone, and this
integration published it as `integer`: on 20;26;Alecto V4;ID=5b99;
TEMP=0128;HUM=37; the 37% was recorded correctly but showed up without
its droplet. Temperature was unaffected because it already used
`decimal`.

Auditing the whole table against the Gladys front turned up one more
case nobody had hit yet: `distance-sensor` is also `decimal`-only. Both
now use SENSOR.DECIMAL. RFLink sends whole numbers either way, so no
decoded value changes; the feature keys are untouched, so the history
survives (Gladys offers an Update on the affected devices).

The pairs in use are pinned by CATEGORY_TYPE_CONTRACT in
test/featureMap.test.js and its counterpart in test/roles.test.js:
nothing in this repository can reach a running Gladys, so adding a field
or a role has to be a deliberate edit to those lists, verified against
front/src/utils/consts.js upstream.

+++ src/devices/featureMap.js
@@ -129,7 +129,13 @@ export const RFLINK_FIELDS = {
     key: 'humidity',
     name: 'Humidity',
     category: CATEGORIES.HUMIDITY_SENSOR,
-    type: TYPES.SENSOR.INTEGER,
+    // DECIMAL even though RFLink only ever sends whole percents: Gladys
+    // resolves a feature's icon and its translated label by the exact
+    // (category, type) pair, and `humidity-sensor` is only registered with
+    // `decimal`. With `integer` the value still arrives, but the droplet icon
+    // falls back to a generic slider and the label to the raw feature name.
+    // See CATEGORY_TYPE_CONTRACT in test/featureMap.test.js.
+    type: TYPES.SENSOR.DECIMAL,
     unit: UNITS.PERCENT,
     min: 0,
     max: 100,
@@ -378,7 +384,8 @@ export const RFLINK_FIELDS = {
     key: 'distance',
     name: 'Distance',
     category: CATEGORIES.DISTANCE_SENSOR,
-    type: TYPES.SENSOR.INTEGER,
+    // Same reason as HUM: `distance-sensor` is only registered with `decimal`.
+    type: TYPES.SENSOR.DECIMAL,
     unit: UNITS.CM,
     min: 0,
     max: 10_000,
